Pull canvas creation helper into own file
