+++ tests/helpers/injected.ts
@@ -1,20 +1,8 @@
-import Canvasimo from 'canvasimo';
 import { Injected } from 'cba';
+import { createCanvas } from './canvas';
 
-const createInjected = <P = {}, S = {}>() => {
-  const element = document.createElement('canvas');
-  jest.spyOn(element, 'getContext').mockImplementation((type: string) => {
-    if (type !== '2d') {
-      throw new Error('Invalid attempted to get context other than 2d');
-    }
-
-    return ({
-      // Prevents unsupported warnings
-      setLineDash: jest.fn(),
-      getLineDash: jest.fn(),
-    } as unknown) as CanvasRenderingContext2D;
-  });
-  const canvas = new Canvasimo(element);
+export function createInjected<P = {}, S = {}>() {
+  const { canvas } = createCanvas();
 
   const injected: Injected<P, S> = {
     canvas,
@@ -25,6 +13,4 @@ const createInjected = <P = {}, S = {}>() => {
   };
 
   return injected;
-};
-
-export { createInjected };
+}
